+++ src/lvl_script.h
@@ -75,9 +75,9 @@ struct TunnellerTrigger {
   unsigned char flags;
   unsigned short condit_idx;
   unsigned char plyr_idx;
-  unsigned long location;
-  unsigned long heading; // originally was 'target'
-  long carried_gold;
+  uint32_t location;
+  uint32_t heading; // originally was 'target'
+  int32_t carried_gold;
   CrtrExpLevel exp_level;
   char party_id;
 };
@@ -94,7 +94,7 @@ struct PartyTrigger {
   union
   {
       TbMapLocation location;
-      unsigned long countdown;
+      uint32_t countdown;
   };
   char spawn_type;
   CrtrExpLevel exp_level;
@@ -124,9 +124,9 @@ struct ScriptValue {
     char chars[32];
     short shorts[16];
     unsigned short ushorts[16];
-    long longs[8];
+    int32_t longs[8];
     long long longlongs[4];
-    unsigned long ulongs[8];
+    uint32_t ulongs[8];
     unsigned long long ulonglongs[4];
   };
 };
@@ -138,7 +138,7 @@ struct Condition {
   unsigned char variabl_type;
   unsigned short variabl_idx;
   unsigned char operation;
-  unsigned long rvalue;
+  uint32_t rvalue;
   unsigned char plyr_range_right;
   unsigned char variabl_type_right;
   unsigned short variabl_idx_right;
@@ -167,23 +167,23 @@ struct ScriptFxLine
 
 struct LevelScript {
     struct TunnellerTrigger tunneller_triggers[TUNNELLER_TRIGGERS_COUNT];
-    unsigned long tunneller_triggers_num;
+    uint32_t tunneller_triggers_num;
     struct PartyTrigger party_triggers[PARTY_TRIGGERS_COUNT];
-    unsigned long party_triggers_num;
+    uint32_t party_triggers_num;
     struct ScriptValue values[SCRIPT_VALUES_COUNT];
-    unsigned long values_num;
+    uint32_t values_num;
     struct Condition conditions[CONDITIONS_COUNT];
-    unsigned long conditions_num;
+    uint32_t conditions_num;
     struct Party creature_partys[CREATURE_PARTYS_COUNT];
-    unsigned long creature_partys_num;
+    uint32_t creature_partys_num;
     unsigned short win_conditions[WIN_CONDITIONS_COUNT];
-    unsigned long win_conditions_num;
+    uint32_t win_conditions_num;
     unsigned short lose_conditions[WIN_CONDITIONS_COUNT];
-    unsigned long lose_conditions_num;
+    uint32_t lose_conditions_num;
 
     // Store strings used at level here
     char strings[2048];
-    long next_string_offset;
+    int32_t next_string_offset;
 };
 
 /******************************************************************************/
